+++ ebl/fragmentarium/web/fragment_date.py
@@ -6,7 +6,7 @@ from marshmallow import ValidationError
 from ebl.fragmentarium.application.fragment_updater import FragmentUpdater
 from ebl.fragmentarium.web.dtos import create_response_dto, parse_museum_number
 from ebl.users.web.require_scope import require_scope
-from ebl.fragmentarium.domain.written_date import DateSchema
+from ebl.fragmentarium.domain.date import DateSchema
 
 
 class FragmentDateResource:
